requirements.txt
streamlit==1.53.0
pandas==2.0.3
gspread==5.7.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
python-dateutil==2.8.2